foreman_password_prompt - update engine to support before_filter
Note: the foreman core has a 'welcome' filter on the application controller.  That
filter can result in a render of a welcome page.  If that occurs, this filter
will never get executed; therefore, either need to disable that filter, override it,
or find a way to inject this filter before it.